resolve float32/64 discrepency

+++ srcs/go/kungfu/kungfu.go
@@ -328,12 +328,12 @@ func parseDelayConfigFile() map[float32]Delay {
 
 func parseDelayFromRow(args []string) Delay {
 	var delayArgs []float32
-	for _, f := range args {
+	for _, f := range args { 
 		j, err := strconv.ParseFloat(f,32)
 		if err != nil {
 			return Delay{1, 0, 0.0}
 		}
-		delayArgs = append(delayArgs, j)
+		delayArgs = append(delayArgs, float32(j))
 	}
 	//output should be iter int, machine int, delay float32
 	delay := Delay{int(delayArgs[0]), int(delayArgs[1]), delayArgs[2]}
